+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased (v0.11.3)
+* Add sample-encode support for computing both XPSNR **and** VMAF using `--xpsnr --do-vmaf`.
+
 # v0.11.2
 * Add arg `--xpsnr-pix-format` for setting an explicit pixel format to use in XPSNR analysis.
   This can be used to workaround some issues with 10-bit video, by setting `--xpsnr-pix-format yuv420p`.